Remove program.rs file re-write and move logic to root - use codama visitor for line re-writes
