--- rust-sdk/client/postprocess-codama.js
@@ -1,54 +1,0 @@
-// Post-processes Codama-generated Rust files to make the program ID
-// runtime-selectable instead of a hardcoded `crate::WHIRLPOOL_ID` const.
-//
-// 1. Leaves `src/generated/programs.rs` alone — that file is hand-maintained
-//    (regenerated by codama as a const, then re-applied by this script).
-// 2. Rewrites `program_id: crate::WHIRLPOOL_ID,` -> `program_id: crate::current_whirlpool_id(),`
-//    inside every generated instruction builder so the dispatch is dynamic.
-
-import { readFileSync, readdirSync, writeFileSync } from "fs";
-import { join } from "path";
-
-const PROGRAMS_FILE = "./src/generated/programs.rs";
-// Lives outside `src/generated/` because the codama renderer wipes files in
-// that directory that aren't part of its output set.
-const RUNTIME_PROGRAMS_FILE = "./programs.runtime.rs";
-
-const REWRITES = [
-  // Instruction builders embed the program ID into the emitted Instruction.
-  {
-    dir: "./src/generated/instructions",
-    from: "program_id: crate::WHIRLPOOL_ID,",
-    to: "program_id: crate::current_whirlpool_id(),",
-  },
-  // Account `anchor_lang::Owner` impls bake in the expected owner program.
-  {
-    dir: "./src/generated/accounts",
-    from: "fn owner() -> Pubkey {\n        crate::WHIRLPOOL_ID\n      }",
-    to: "fn owner() -> Pubkey {\n        crate::current_whirlpool_id()\n      }",
-  },
-];
-
-for (const { dir, from, to } of REWRITES) {
-  let rewritten = 0;
-  for (const name of readdirSync(dir)) {
-    if (!name.endsWith(".rs")) continue;
-    const path = join(dir, name);
-    const before = readFileSync(path, "utf8");
-    if (!before.includes(from)) continue;
-    writeFileSync(path, before.split(from).join(to));
-    rewritten += 1;
-  }
-  console.info(`postprocess-codama: rewrote ${rewritten} files in ${dir}`);
-}
-
-// Replace the codama-emitted const-only programs.rs with the runtime-selectable
-// variant maintained at programs.runtime.rs. Skipped silently if the runtime
-// template is not present (e.g. during the initial bootstrap).
-try {
-  const runtime = readFileSync(RUNTIME_PROGRAMS_FILE, "utf8");
-  writeFileSync(PROGRAMS_FILE, runtime);
-  console.info("postprocess-codama: applied runtime programs.rs");
-} catch (err) {
-  if (err.code !== "ENOENT") throw err;
-}
